requirements.txt
asgiref==3.3.1
astroid==2.4.2
confusable-homoglyphs==3.2.0
Django==3.1.5
django-extra-views==0.13.0
django-recaptcha==2.0.6
django-registration==3.1.1
django-resized==0.3.11
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
psycopg2==2.8.6
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2020.5
rope==0.18.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
